backend-api/requirements.txt
# Core FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-multipart==0.0.6

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0